Improvements/fixes
