+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/model/GwtChannelRecord.java
@@ -11,8 +11,6 @@ package org.eclipse.kura.web.shared.model;
 
 import java.io.Serializable;
 
-import org.eclipse.kura.web.client.util.KuraBaseModel;
-
 public class GwtChannelRecord extends KuraBaseModel implements Serializable {
 
     /** Serialization UUID */
