+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/model/GwtConfigComponent.java
@@ -13,8 +13,6 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.eclipse.kura.web.client.util.KuraBaseModel;
-
 public class GwtConfigComponent extends KuraBaseModel implements Serializable {
 
     private static final long serialVersionUID = -6388356998309026758L;
