+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/model/GwtFirewallPortForwardEntry.java
@@ -13,8 +13,6 @@ package org.eclipse.kura.web.shared.model;
 
 import java.io.Serializable;
 
-import org.eclipse.kura.web.client.util.KuraBaseModel;
-
 public class GwtFirewallPortForwardEntry extends KuraBaseModel implements Serializable {
 
     /**
