+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/model/GwtWifiConfig.java
@@ -14,7 +14,7 @@ package org.eclipse.kura.web.shared.model;
 import java.io.Serializable;
 import java.util.List;
 
-import org.eclipse.kura.web.client.util.GwtSafeHtmlUtils;
+import org.eclipse.kura.web.shared.GwtSafeHtmlUtils;
 
 public class GwtWifiConfig extends GwtBaseModel implements Serializable {
