+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/model/GwtWifiHotspotEntry.java
@@ -12,8 +12,6 @@
 package org.eclipse.kura.web.shared.model;
 
 import java.io.Serializable;
-
-import org.eclipse.kura.web.client.util.KuraBaseModel;
 import java.util.logging.Level;
 import java.util.logging.Logger;
