+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/model/KuraBaseModel.java
@@ -9,9 +9,9 @@
  * Contributors:
  *     Eurotech
  *******************************************************************************/
-package org.eclipse.kura.web.client.util;
+package org.eclipse.kura.web.shared.model;
 
-import org.eclipse.kura.web.shared.model.GwtBaseModel;
+import org.eclipse.kura.web.shared.GwtSafeHtmlUtils;
 
 public class KuraBaseModel extends GwtBaseModel {
