+++ kura/org.eclipse.kura.web2/src/main/resources/org/eclipse/kura/web/js/dnd_helper.js
