+++ kura/org.eclipse.kura.web2/src/main/resources/org/eclipse/kura/web/js/joint.shapes.devs.js
