+++ kura/org.eclipse.kura.web2/src/main/resources/org/eclipse/kura/web/js/wires_composer.js
