+++ kura/org.eclipse.kura.web2/src/main/webapp/auth.html
@@ -12,7 +12,8 @@
     <!--                                                               -->
     <!-- Consider inlining CSS to reduce the number of requested files -->
     <!--                                                               -->
-    <link type="text/css" rel="stylesheet" href="/admin/denali.css">
+    <script type="text/javascript" src="/admin/login.all-min.js"></script>
+    <link type="text/css" rel="stylesheet" href="/admin/denali.all-min.css">
     <link type="text/css" rel="stylesheet" href="/admin/denali/skin/skin.css?v=1"/>
 	<!--                                           -->
     <!-- Any title is fine                         -->
@@ -29,8 +30,8 @@
 
     <!-- Set viewport scale for mobile devices -->
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    
-    
+
+
   </head>
 
   <!--                                           -->
