+++ kura/org.eclipse.kura.web2/src/main/webapp/denali.html
@@ -12,16 +12,8 @@
     <!--                                                               -->
     <!-- Consider inlining CSS to reduce the number of requested files -->
     <!--                                                               -->
-    <link type="text/css" rel="stylesheet" href="joint.min.css">
-    <script type="text/javascript" src="/admin/jquery.min.js"></script>
-    <script type="text/javascript" src="/admin/lodash.min.js"></script>
-    <script type="text/javascript" src="/admin/backbone.min.js"></script>
-    <script type="text/javascript" src="/admin/joint.min.js"></script>
-    <script type="text/javascript" src="/admin/joint.shapes.devs.js"></script>
-    <script type="text/javascript" src="/admin/dnd_helper.js"></script>
-    <script type="text/javascript" src="/admin/wires_composer.js"></script>
-    <script type="text/javascript" src="/admin/wires_blink.js"></script>
-    <link type="text/css" rel="stylesheet" href="/admin/denali.css">
+    <link type="text/css" rel="stylesheet" href="/admin/denali.all-min.css">
+    <script type="text/javascript" src="/admin/denali.all-min.js"></script>
     <link type="text/css" rel="stylesheet" href="/admin/denali/skin/skin.css?v=1"/>
 	<!--                                           -->
     <!-- Any title is fine                         -->
